Relocate files to proposals root
