+++ __e2e__/proposals/joinSpaceViaProposals.spec.ts
@@ -6,7 +6,7 @@ import { test } from '__e2e__/utils/test';
 import { baseUrl } from 'testing/mockApiCall';
 import { generateProposalWorkflow } from 'testing/utils/proposals';
 
-import { generateUser, generateUserAndSpace, loginBrowserUser, logoutBrowserUser } from '../../utils/mocks';
+import { generateUser, generateUserAndSpace, loginBrowserUser, logoutBrowserUser } from '../utils/mocks';
 
 test('Create a proposal from a linked proposal template /member', async ({ page, proposalPage, documentPage }) => {
   const { user: admin, space } = await generateUserAndSpace({ isAdmin: true });
